server/requirements.txt
fastapi
uvicorn[standard]
python-socketio
anthropic
